DMRECON: rename files for consistency
* use lower case with underscores
* use .cc instead of .cpp file extension
